% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startifyR.R
\name{startify}
\alias{startify}
\title{Create a random quote with an ascii character}
\usage{
startify(width = 60)
}
\arguments{
\item{width}{width of the ascii art, should not exceed terminal width}
}
\description{
Create a random quote with an ascii character
}
